* commit '5ae5f8f0de6651472534ce3b70b0cf65f1339503': (28 commits)
  Revert "Calculate sorted motion supporter list (#2805)" (#2809)
  Fix force reload not working (#2808)
  Calculate sorted motion supporter list (#2805)
  changed all instances of factory defaults to default settings (#2804)
  Fix hidden agenda projection in autopilot (#2807)
  Fix scrolling to change recommandation (#2801)
  Fixed motion pdf filename (#2798)
  Reverts PR #2375 (#2797)
  re-add speaker even if PoO (#2794)
  Client performance improvements (#2786)
  Fix speaker column in XLSX export (#2796)
  Make sort lists not update during drag/drop (#2772)
  Fix projector preview (#2777)
  Fix supporter edit in motion edit (#2790)
  Revert "Fix single option chart table style (#2774)" (#2793)
  Fix vote delegation (#2788)
  Make motion multiselect csv/pdf export sort by current sorting (#2783)
  Add 'No changes at the text' description for amendments without changes (#2784)
  Use BroadcastChannel in shared worker for broadcasting (#2781)
  Fix linter warnings (#2754)
  ...

+++ client/src/app/site/services/data-store-update-manager.service.ts
@@ -176,18 +176,22 @@ export class DataStoreUpdateManagerService {
         const repositories = this.mapperService.getAllRepositories();
 
         // Phase 1: deleting and creating of view models (in this order)
-        repositories.forEach(repo => {
+        for (const repo of repositories) {
             const deletedModelIds = slot.getDeletedModelIdsForCollection(repo.collection);
-            repo.deleteModels(deletedModelIds);
+            if (deletedModelIds?.length) {
+                repo.deleteModels(deletedModelIds);
+            }
 
             const changedModelIds = slot.getChangedModelIdsForCollection(repo.collection);
-            repo.changedModels(changedModelIds, changedModels[repo.COLLECTION]);
-        });
+            if (changedModelIds?.length) {
+                repo.changedModels(changedModelIds, changedModels[repo.COLLECTION]);
+            }
+        }
 
         // Phase 2: updating all repositories
-        repositories.forEach(repo => {
+        for (const repo of repositories) {
             repo.commitUpdate(slot.getAllModelsIdsForCollection(repo.collection));
-        });
+        }
 
         slot.DS.triggerModifiedObservable();
